telegram_bot/telegram_bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
Pillow==10.2.0
numpy==1.26.3

